refactor: board 패키지 생성
