+++ src/main/java/domain/Board.java
@@ -1,6 +1,6 @@
 package domain;
 
-import domain.position.Position;
+import domain.board.Position;
 
 import java.util.Collections;
 import java.util.HashSet;
