+++ src/test/java/domain/position/RankTest.java
@@ -1,5 +1,6 @@
 package domain.position;
 
+import domain.board.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
