+++ src/test/java/domain/strategy/PawnMoveStrategyTest.java
@@ -1,9 +1,9 @@
 package domain.strategy;
 
 import domain.TeamColor;
-import domain.position.File;
-import domain.position.Position;
-import domain.position.Rank;
+import domain.board.File;
+import domain.board.Position;
+import domain.board.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
